Implement `build_alias`
Copy of the functionality in W. Caleb McDaniel's Bash script implemented
in Rust:

    echo "${MESSAGE}" | grep ^"From: " | sed s/[\,\"\']//g | awk '{$1=""; if (NF == 3) {print "alias" $0;} else if (NF == 2) {print "alias" $0 $0;} else if (NF > 3) {print "alias", tolower($(NF-1))"-"tolower($2) $0;}}'

Doesn't currently include the [,"'] filtering.

Added a couple of new test cases to cover other address formats.

Not sure if I should use `if..else if` here or a `match`. I feel like I
should return an empty string at the end if we can't build an alias
line.