Alias#to_string: Don't format `name` if it's empty
When we get a "From: " address that doesn't contain a name, the
resulting `Alias` object created by `Alias::new` will have an empty
name.

When `to_string` is called on these kinds of objects, an extra space
appears between the alias and the email address because of how the
format string is constructed.

This caused a test failure for the `new_alias_with_only_email` test
method in fc05eb9e2f111642dbec093218b7184f25740b90. Our change here
fixes and passes this test.