Rename is_alias_in_file to find_alias_in_file
I think we should return a list of matched aliases from this function
instead of a simple boolean value. As such, we'll start by renaming the
function to something that makes sense in that context.

We now need to modify the function and return type in order to provide
that information.