Add test find_alias_in_file_email_already_exists
When the email of the alias we're looking is already in the file, expect
an `AliasSearchError::EmailExists` error.

Oh man, this was a tough one. After much searching, finally figured out
how to implement the `PartialEq` trait for my error type so that we
could actually test it.

Many thanks to @peterbudai for an example in the 'redux' project of how
to do this
(https://github.com/peterbudai/redux/blob/ef5d47a0a64cef9fa9e1e9c6f21badc46fa283fc/src/lib.rs):

    #[cfg(test)]
    impl PartialEq<Error> for Error {
        fn eq(&self, other: &Error) -> bool {
           match *self {
               Error::Eof => match *other { Error::Eof => true, _ => false },
               Error::InvalidInput => match *other { Error::InvalidInput => true, _ => false },
               Error::IoError(_)  => match *other { Error::IoError(_) => true, _ => false },
           }
        }
    }

With that example, I was able to correctly build an equality function to
get past my compiler errors which complained that

    an implementation of `std::cmp::PartialEq` might be missing for `std::io::error::Error`

when I tried to `#[derive(PartialEq)]` on my `AliasSearchError` type.