Add test for `find_alias_in_file` Ok case
New test for when alias matches are found in the given file. We want to
match all aliases that don't have the searched email and start with the
searched alias string.

This gives us a list of aliases. We'll then be able to use this list to
determine the highest-ranking id of the searched alias and append
${id}+1 to it to create our new alias.