find_alias_in_file: Return list of aliases not alias lines
My next step is to get the list produced by this function and use it to
build a new alias of `#{alias}-#{id + 1}`.

In trying to figure out how best to do that, I realised that it would be
easier to do if I had actual aliases to work with instead of full Mutt
alias lines that I'd have to parse (again).

Update our function to give us a list of aliases instead of full alias
lines.