test.sh: Initial stab at an integration test
Would like to modify the program to take a `--file` attribute so that we
can supply a file on the command line to do an integration test with.

The idea is that we pipe an email to the program and check that the
right alias was added to our aliases file.