Add integration test
Use the TAP harness with `prove` to give ourselves an integration test
of the executable.

This checks that when passing an email to it over STDIN, the proper
alias gets added to the aliases file.

For the moment, we're using the temporary `./testaliases` file. In the
future we'll want to update our program and test code to use something
like `./testdata/tmp` or something like that.