`write_to_file` test: Extract append lines to closure
We want to be able to write a test that ensures aliases can be written
to the file even if they don't already exist.

To do this, I want to be able to reuse the code in the helper function.
We can't use the code that appends to the file because this is relevant
only to the `alias_write_to_file_must_write_given_alias_to_file` test
(which I just realised should actually be renamed to something more
specific).

In order to run this code without requiring it to be in the helper
function, extract it to a closure that gets passed to the helper.

We need to pass `alias` into the function explicitly in order to use it
otherwise we get an error on an immutable borrow.