Write alias even if we haven't found a matching one
If we get an `AliasSearchError::NotFound` error we still want to write
the alias to the file. After all, it's an email we haven't encountered
yet so we should store an alias for it.

If the email exists, we can do what we were doing before and change the
current alias to be unique.

Rename our "append" `write_to_file` test method to be a little more
clear about its purpose.

Create a new `write_to_file` test function that tests the case we're
adding here.

Note that having these two tests together seems to cause a race
condition because they're both using the same temporary test file. If we
run tests with `RUST_TEST_THREADS=1 cargo test` then they pass. Will see
about fixing that next.